fix: pre-populate search bar with query
Previous behaviour was that the last query would then be used as the input's placeholder
This was slightly confusing, as a query, e.g. 'coral', would seem to persist and be undeletable

`#resources_search_placeholder` now returns a non-dynamic string: 'Search resources by label/name', so the placeholder will always read that text
`#resources_search_initial_value` has now been added to produce the string which is passed to the input element as its value
If there's already a query parameter, then the input element will begin with that as its value. Otherwise, it will begin with an empty string (thereby revealing the placeholder)